pyproject.toml
[project]
name = "ktr-map-display"
version = "0.1.0"
description = "A cli tool to display maps from Killing Time:Resurrected"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.2",
    "omgifol>=0.5.1",
    "pillow>=12.2.0",
    "rich>=15.0.0",
    "rich-click>=1.9.7",
]

[project.scripts]
ktr-display = "main:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.12.0",
]
